refactor: cleanup file indirection
I put file rename in another commit to simplify the diff.
